requirements.txt
python-dotenv
requests
pandas
fpdf2
openpyxl
matplotlib